+++ apps/design-system-docs/docusaurus.config.ts
@@ -14,7 +14,7 @@ const config: Config = {
   //   v4: true, // Improve compatibility with the upcoming Docusaurus v4
   // },
 
-  url: 'https://equinor.github.io',
+  url: 'https://eds.equinor.com',
   baseUrl: '/',
 
   organizationName: 'equinor',
@@ -57,6 +57,9 @@ const config: Config = {
             './src/css/custom.css',
             // Add additional CSS files here
           ],
+        },
+        sitemap: {
+          lastmod: 'datetime',
         },
       } satisfies Preset.Options,
     ],