+++ packages/eds-core-react/src/components/Button/tokens/outlined.ts
@@ -1,6 +1,7 @@
 import { tokens } from '@equinor/eds-tokens'
 import { mergeDeepRight } from 'ramda'
 import { button } from './button'
+import { ButtonToken } from '../Button.types'
 
 const {
   colors: {
@@ -50,7 +51,7 @@ export const primary = mergeDeepRight(button, {
       background: 'transparent',
     },
   },
-})
+}) as ButtonToken
 
 export const secondary = mergeDeepRight(primary, {
   typography: {
@@ -70,7 +71,7 @@ export const secondary = mergeDeepRight(primary, {
       },
     },
   },
-})
+}) as ButtonToken
 
 export const danger = mergeDeepRight(primary, {
   typography: {
@@ -90,4 +91,4 @@ export const danger = mergeDeepRight(primary, {
       },
     },
   },
-})
+}) as ButtonToken