+++ packages/eds-core-react/src/components/next/CHANGELOG.md
@@ -4,6 +4,41 @@ All notable changes to EDS 2.0 beta components (`@equinor/eds-core-react/next`) 
 
 These are experimental components available under the `/next` entry point. They follow semantic versioning with a `beta` prerelease tag.
 
+## [3.0.0-beta.1](https://github.com/equinor/design-system/compare/eds-core-react-next@v2.8.0-beta.1...eds-core-react-next@v3.0.0-beta.1) (2026-07-20)
+
+
+### ⚠ BREAKING CHANGES
+
+* consumers targeting these class names directly in their own CSS must update to the new flat names.
+
+### 🐛 Fixed
+
+* convert Banner BEM classes to flat class names ([#5114](https://github.com/equinor/design-system/issues/5114)) ([778515c](https://github.com/equinor/design-system/commit/778515ca9e2eb6144fcf96f57b483d6afdaae25d))
+* convert Field, Switch, Radio, Checkbox BEM classes to flat class names ([#5115](https://github.com/equinor/design-system/issues/5115)) ([84906f3](https://github.com/equinor/design-system/commit/84906f3932d6049190ca91859c7b9a03968ab7d3))
+
+## [2.8.0-beta.1](https://github.com/equinor/design-system/compare/eds-core-react-next@v2.7.0-beta.1...eds-core-react-next@v2.8.0-beta.1) (2026-06-23)
+
+
+### ✨ Added
+
+* Badge EDS 2.0 ([#4999](https://github.com/equinor/design-system/issues/4999)) ([2235377](https://github.com/equinor/design-system/commit/2235377b318cb4c75712d6b2380d16caa3aef721))
+* Dialog EDS 2.0 ([#4956](https://github.com/equinor/design-system/issues/4956)) ([e2f6df2](https://github.com/equinor/design-system/commit/e2f6df2f62991d3f5a0205b416ffcf5d880a3d17))
+* Select EDS 2.0 ([#4918](https://github.com/equinor/design-system/issues/4918)) ([0be44ad](https://github.com/equinor/design-system/commit/0be44adfbf0a45ad27b6b6b5f3a0cfb7337c4ec9))
+
+
+### 🐛 Fixed
+
+* correct TextArea helper message spacing in both density modes ([#4986](https://github.com/equinor/design-system/issues/4986)) ([f1a3b4d](https://github.com/equinor/design-system/commit/f1a3b4dcab016f4236e0d10ffcf63af4e25d6486))
+* remove box-shadow from Switch handle ([#4980](https://github.com/equinor/design-system/issues/4980)) ([2f0dd96](https://github.com/equinor/design-system/commit/2f0dd96d685107e5d770bbfb00b8dde8cf80869a)), closes [#4838](https://github.com/equinor/design-system/issues/4838)
+* replace styled-components Progress with spinning icon in Autocomplete ([#4982](https://github.com/equinor/design-system/issues/4982)) ([9581f89](https://github.com/equinor/design-system/commit/9581f892ef28d8d4fa1195ef05818c5fadf34f4c))
+* rely on parent flex gap for Field helper-message spacing ([#5080](https://github.com/equinor/design-system/issues/5080)) ([8a18fb6](https://github.com/equinor/design-system/commit/8a18fb697aaeab3bdab6f14bd3a47160377b58a7))
+* replace styled-components Tooltip with next/Tooltip in TextField and TextArea ([#4981](https://github.com/equinor/design-system/issues/4981)) ([e3ea724](https://github.com/equinor/design-system/commit/e3ea724eb99e9a3511686c07a2d7b9232f38cb10))
+
+
+### ♻️ Refactoring
+
+* migrate Banner.Message and Chip label off TypographyNext ([#4984](https://github.com/equinor/design-system/issues/4984)) ([ea19590](https://github.com/equinor/design-system/commit/ea195904f5321e89930f3d0247025ac4deeb9283)), closes [#4835](https://github.com/equinor/design-system/issues/4835) [#4836](https://github.com/equinor/design-system/issues/4836)
+
 ## [2.7.0-beta.1](https://github.com/equinor/design-system/compare/eds-core-react-next@v2.6.0-beta.1...eds-core-react-next@v2.7.0-beta.1) (2026-05-27)